fix favoris view sale/search post . browse SearchPost non connecté . pull to refresh messages . redirection sale post depuis conv .

+++ src/app/store/sale/view-sale-post/view-sale-post.component.ts
@@ -1,15 +1,16 @@
-import {Component, OnInit, Renderer2} from '@angular/core';
+import {Component, Input, OnInit, Renderer2} from '@angular/core';
 import {ActivatedRoute, Router} from "@angular/router";
 import { Location } from '@angular/common'
 import {combineLatest, concat, filter, map, Observable, of, ReplaySubject, Subject, switchMap, take, tap} from "rxjs";
 import {PokemonItemReferenceModel} from "../../../core/models/pokemon-item-reference.model";
 import {SalePostModel} from "../../../core/models/sale-post.model";
 import {UserService} from "../../../core/services/UserService/user.service";
-import { ModalController } from '@ionic/angular';
+import {ActionSheetController, ModalController} from '@ionic/angular';
 import { FullScreenImageComponent } from 'src/app/core/components/full-screen-image/full-screen-image.component';
 import { FullScreenImageSliderComponent } from 'src/app/core/components/full-screen-image-slider/full-screen-image-slider.component';
 import {PaymentComponent} from "../payment/payment.component";
 import {SalePostService} from "../services/sale-post.service";
+import {LikedSalePostService} from "../../../core/services/LikedSalePostService/liked-sale-post.service";
 
 @Component({
   selector: 'app-view-sale-post',
@@ -19,16 +20,20 @@ import {SalePostService} from "../services/sale-post.service";
 export class ViewSalePostComponent {
   SalePost$:Observable<SalePostModel | null> = of(new SalePostModel())
   UserSalePost$:Observable<SalePostModel[] | null> = of([])
-  loading: boolean = true;
+  @Input() paramSearchPostId:string = '';
+
   constructor(private salePostService:SalePostService,
               private route:ActivatedRoute,
               private modalCtrl: ModalController,
               private userService:UserService,
               private router:Router,
               private renderer: Renderer2,
-              private location: Location) {
+              private location: Location,
+              private  likedSalePostService:LikedSalePostService,
+              private actionSheetCtrl: ActionSheetController) {
   }
-
+  loading: boolean = true;
+  loadingLike: boolean = false;
   isDisabled:boolean = true;
   isOwner: boolean = false;
   searchPostId: string = this.route.snapshot.params['id'];
@@ -37,12 +42,17 @@ export class ViewSalePostComponent {
   private userId$ = new ReplaySubject<number>();
 
   ngOnInit() {
+
+        if(this.searchPostId === undefined){
+          this.searchPostId = this.paramSearchPostId
+        }
+
     this.userId$.pipe(take(1)).subscribe(userId => {
       this.DefineOwner(userId);
       this.initUserSalePost$();
     });
 
-    this.SalePost$ = this.salePostService.getSingleSalePost(this.route.snapshot.params['id']).pipe(
+    this.SalePost$ = this.salePostService.getSingleSalePost(this.searchPostId).pipe(
       filter((value) => value !== null),
       tap(_ => this.loading = false),
       map(response => response.body),
@@ -117,14 +127,68 @@ export class ViewSalePostComponent {
     }
   }
 
-  deleteSalePost(id: string) {
-    this.salePostService.deleteSalePost(id).subscribe(() => {
-      this.location.back()
-    })
+  async deleteSalePost(id: string) {
+    const confirm = await this.canDismiss();
+    if (confirm) {
+      this.salePostService.deleteSalePost(id).subscribe(() => {
+        this.location.back()
+      })
+    }
   }
 
   redirectToOfferPage(idSalePost: string){
     this.router.navigateByUrl(`/tabs/store/sale/offer/${idSalePost}` );
   }
+
+  LikeSalePost(id: string, event: MouseEvent) {
+    this.loadingLike = true;
+    event.stopPropagation();
+    this.likedSalePostService.LikeSalePost(id).pipe(
+      switchMap(_ => this.salePostService.getSingleSalePost(this.route.snapshot.params['id'])),
+      filter((value) => value !== null),
+      tap(_ => this.loadingLike = false),
+      map(response => response.body),
+      tap(resp => this.userId$.next(resp!.userId)),
+    ).subscribe(salePost => this.SalePost$ = of(salePost));
+  }
+
+  UnlikeSalePost(id: string, event: MouseEvent) {
+    this.loadingLike = true;
+    event.stopPropagation();
+    this.likedSalePostService.UnLikeSalePost(id).pipe(
+      switchMap(_ => this.salePostService.getSingleSalePost(this.route.snapshot.params['id'])),
+      filter((value) => value !== null),
+      tap(_ => this.loadingLike = false),
+      map(response => response.body),
+      tap(resp => this.userId$.next(resp!.userId)),
+    ).subscribe(salePost => this.SalePost$ = of(salePost));
+  }
+
+  async dismiss() {
+    await this.modalCtrl.dismiss();
+  }
+
+  canDismiss = async () => {
+    const actionSheet = await this.actionSheetCtrl.create({
+      mode: 'ios',
+      header: 'Êtes vous sûr ?',
+      buttons: [
+        {
+          text: 'Oui',
+          role: 'confirm',
+        },
+        {
+          text: 'Non',
+          role: 'cancel',
+        },
+      ],
+    });
+
+    await actionSheet.present();
+
+    const { role } = await actionSheet.onWillDismiss();
+
+    return role === 'confirm';
+  };
 
 }
