+++ src/app/store/search/view-search-post/view-search-post.component.ts
@@ -6,6 +6,8 @@ import {filter, map, Observable, of, ReplaySubject, switchMap, take, tap} from "
 import {HttpResponse} from "@angular/common/http";
 import {UserService} from "../../../core/services/UserService/user.service";
 import {Location} from "@angular/common";
+import {ActionSheetController, ModalController} from "@ionic/angular";
+import {LikedSearchPostService} from "../../../core/services/LikedSearchPostService/liked-search-post.service";
 
 @Component({
   selector: 'app-view-search-post',
@@ -17,12 +19,16 @@ export class ViewSearchPostComponent {
   SearchPost$:Observable<SearchPostModel | null> = of(new SearchPostModel())
   UserSearchPost$:Observable<SearchPostModel[] | null> = of([])
   loading: boolean = true;
+  public loadingLike: boolean = false;
   constructor(private searchPostService:SearchPostService,
               private route:ActivatedRoute,
               private userService:UserService,
               private router:Router,
               private renderer: Renderer2,
-              private location: Location) {
+              private location: Location,
+              private  likedSearchPostService:LikedSearchPostService,
+              private actionSheetCtrl: ActionSheetController,
+              private modalCtrl: ModalController,) {
 
   }
 
@@ -49,7 +55,7 @@ export class ViewSearchPostComponent {
 
     this.userId$.pipe(take(1)).subscribe(userId => {
       this.DefineOwner(userId);
-      this.initUserSalePost$();
+      this.initUserSearchPost$();
     });
 
     this.SearchPost$ = this.searchPostService.getSingleSearchPost(this.route.snapshot.params['id']).pipe(
@@ -60,7 +66,7 @@ export class ViewSearchPostComponent {
     );
   }
 
-  initUserSalePost$() {
+  initUserSearchPost$() {
     this.UserSearchPost$ = this.userId$.pipe(
       switchMap(userId => this.searchPostService.getSomeSearchPostForUser(userId, 5)),
       filter((value) => value !== null),
@@ -93,9 +99,63 @@ export class ViewSearchPostComponent {
     }
   }
 
-  deleteSearchPost(id: string) {
-    this.searchPostService.deleteSearchPost(id).subscribe(() => {
-      this.location.back()
-    })
+  async deleteSearchPost(id: string) {
+    const confirm = await this.canDismiss();
+    if (confirm) {
+      this.searchPostService.deleteSearchPost(id).subscribe(() => {
+        this.location.back()
+      })
+    }
   }
+
+  LikeSearchPost(id: string, event: MouseEvent) {
+    this.loadingLike = true;
+    event.stopPropagation();
+    this.likedSearchPostService.LikeSearchPost(id).pipe(
+      switchMap(_ => this.searchPostService.getSingleSearchPost(this.route.snapshot.params['id'])),
+      filter((value) => value !== null),
+      tap(_ => this.loadingLike = false),
+      map(response => response.body),
+      tap(resp => this.userId$.next(resp!.userId)),
+    ).subscribe(searchPost => this.SearchPost$ = of(searchPost));
+  }
+
+  UnlikeSearchPost(id: string, event: MouseEvent) {
+    this.loadingLike = true;
+    event.stopPropagation();
+    this.likedSearchPostService.UnLikeSearchPost(id).pipe(
+      switchMap(_ => this.searchPostService.getSingleSearchPost(this.route.snapshot.params['id'])),
+      filter((value) => value !== null),
+      tap(_ => this.loadingLike = false),
+      map(response => response.body),
+      tap(resp => this.userId$.next(resp!.userId)),
+    ).subscribe(searchPost => this.SearchPost$ = of(searchPost));
+  }
+
+  async dismiss() {
+    await this.modalCtrl.dismiss();
+  }
+
+  canDismiss = async () => {
+    const actionSheet = await this.actionSheetCtrl.create({
+      mode: 'ios',
+      header: 'Êtes vous sûr ?',
+      buttons: [
+        {
+          text: 'Oui',
+          role: 'confirm',
+        },
+        {
+          text: 'Non',
+          role: 'cancel',
+        },
+      ],
+    });
+
+    await actionSheet.present();
+
+    const { role } = await actionSheet.onWillDismiss();
+
+    return role === 'confirm';
+  };
 }
